+++ README.md
@@ -34,3 +34,4 @@
 * Промежуточный коммит (commit 75)
 * Промежуточный коммит (commit 76)
 * Промежуточный коммит. Очистка кода перед review (commit 77)
+* Реализованы все задачи по функционалу и обмену данными (состояними) через хранилище redux предусмльренные спринтом 8 и проектом month-8/step-2. (commit 78)
