+++ src/components/services/actions/burger-ingredients.js
@@ -1,15 +1,4 @@
 export const ITEM_OPEN = 'ITEM_OPEN';
 export const ITEM_CLOSE = 'ITEM_CLOSE';
-
 export const MODAL_OPEN = 'MODAL_OPEN';
 export const MODAL_CLOSE = 'MODAL_CLOSE';
-
-// экшены перемещения булок и перемещения соусов и начинок
-
-export const BUN_MOVE = 'BUN_MOVE';
-export const SAUCE_FILLING_MOVE = 'SAUCE_FILLING_MOVE';
-export const ELEMENT_REMOVE = 'ELEMENT_REMOVE';
-
-// Обновление массива сортируемых едлементов
-
-export const UPDATE_ARR_ELEMENTS = 'UPDATE_ARR_ELEMENTS';
