+++ src/components/services/reducers/index.js
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 import { getDataReducer,  getOrderReducer} from './api';
-import { itemReducer, madalReducer, ingredientMoveReducer} from './burger-ingredients';
+import { itemReducer, madalReducer} from './burger-ingredients';
+import { ingredientMoveReducer} from './burger-constructor';
 
 export const rootReducer = combineReducers({
   data: getDataReducer,
